requirements.txt
django
psycopg2
netsnmp-py
#libsnmp-dev
#libzmq3-dev
#libczmq-dev
